+++ tests/Unit/Validation.php
@@ -71,12 +71,18 @@ it('can run validation using method is_valid with closure (param)', function () 
         'test1' => 'test',
         'test2' => 'test',
         'test3' => 'test',
+        'test4' => 'test',
+        'test5' => 'test',
+        'test6' => 'test',
+        'test7' => 'test',
     ]);
 
     expect($valid->is_valid(function (ValidPool $pool) {
         $pool->rule('test1')->required();
         $pool('test2')->required();
         $pool->test3->required();
+        $pool->rule('test4', 'test5')->required();
+        $pool('test6', 'test7')->required();
     }))->toBeTrue();
 });
 
@@ -85,6 +91,8 @@ it('can run validation using method is_valid with closure (return)', function ()
         'test1' => 'test',
         'test2' => 'test',
         'test3' => 'test',
+        'test4' => 'test',
+        'test5' => 'test',
     ]);
 
     expect($valid->is_valid(function () {
@@ -92,6 +100,7 @@ it('can run validation using method is_valid with closure (return)', function ()
         $pool->rule('test1')->required();
         $pool('test2')->required();
         $pool->test3->required();
+        $pool->rule('test4', 'test5')->required();
 
         return $pool;
     }))->toBeTrue();
@@ -182,6 +191,10 @@ it('can run filter using method filter_out with closure (param)', function () {
         'test1' => 'test',
         'test2' => ' test ',
         'test3' => 'TEST',
+        'test4' => ' test ',
+        'test5' => ' test ',
+        'test6' => ' test ',
+        'test7' => ' test ',
     ]);
 
     expect(
@@ -189,11 +202,17 @@ it('can run filter using method filter_out with closure (param)', function () {
             $pool->rule('test1')->upper_case();
             $pool->test2->trim();
             $pool('test3')->lower_case();
+            $pool->rule('test4', 'test5')->trim();
+            $pool('test6', 'test7')->trim();
         })
     )->toEqual([
         'test1' => 'TEST',
         'test2' => 'test',
         'test3' => 'test',
+        'test4' => 'test',
+        'test5' => 'test',
+        'test6' => 'test',
+        'test7' => 'test',
     ]);
 });
 
@@ -202,6 +221,10 @@ it('can run filter using method filter_out with closure (return)', function () {
         'test1' => 'test',
         'test2' => ' test ',
         'test3' => 'TEST',
+        'test4' => ' test ',
+        'test5' => ' test ',
+        'test6' => ' test ',
+        'test7' => ' test ',
     ]);
 
     expect(
@@ -210,6 +233,8 @@ it('can run filter using method filter_out with closure (return)', function () {
             $pool->rule('test1')->upper_case();
             $pool->test2->trim();
             $pool('test3')->lower_case();
+            $pool->rule('test4', 'test5')->trim();
+            $pool('test6', 'test7')->trim();
 
             return $pool;
         })
@@ -217,6 +242,10 @@ it('can run filter using method filter_out with closure (return)', function () {
         'test1' => 'TEST',
         'test2' => 'test',
         'test3' => 'test',
+        'test4' => 'test',
+        'test5' => 'test',
+        'test6' => 'test',
+        'test7' => 'test',
     ]);
 });
 
@@ -267,4 +296,33 @@ it('can get error message when valadation is fallen using method validOrError', 
     $valid->field('test')->min_len(5);
 
     expect($valid->validOrError())->toHaveCount(1);
+});
+
+// can add multi validate rule
+it('can add multy field using method field', function () {
+    $valid = new Validator(['test' => 'test', 'test2' => 'test']);
+
+    $valid->field('test', 'test2')->required();
+
+    expect($valid->is_valid())->toBeTrue();
+});
+
+it('can add multy field using method __invoke', function () {
+    $valid = new Validator(['test' => 'test', 'test2' => 'test']);
+
+    $valid('test', 'test2')->required();
+
+    expect($valid->is_valid())->toBeTrue();
+});
+
+// can add multi filter rule
+it('can add multy filter using method filter', function () {
+    $valid = new Validator(['test' => ' test ', 'test2' => ' test ']);
+
+    $valid->field('test', 'test2')->required();
+    $valid->filter('test', 'test2')->trim();
+
+    expect($valid->filter_out())->toMatchArray(
+        ['test' => 'test', 'test2' => 'test']
+    );
 });